# server/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "server API for stock web",
        "title": "Stock-Show Web API",
        "termsOfService": "http://github.com/greetlist/",
        "contact": {
            "name": "Greetlist",
            "url": "http://github.com/greetlist",
            "email": "lrt12250914@outlook.com"
        },
        "version": "1.0"
    },
    "basePath": "/",
    "paths": {
...
        "/api/stock/getRecommandStockPrediction": {
            "post": {
                "description": "Return Recommand Stock Prediction Data",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "Query Recommand Stock Prediction Data",
                "operationId": "getRecommandStockPrediction",
                "parameters": [
                    {
                        "description": "Query Date",
                        "name": "request_json",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.GetRecommandStockRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.GetRecommandStockResponse"
                        }
                    }
                }
            }
        },
        "/api/stock/getTotalMarketIndexData": {
            "post": {
                "description": "Return Strategy Sepcific Computed Data",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "Query Sepcific All Index Computed Data",
                "operationId": "getTotalMarketIndexData",
                "parameters": [
                    {
                        "description": "Query Index ",
                        "name": "request_json",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.GetTotalMarketIndexDataRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.GetTotalMarketIndexDataResponse"
                        }
                    }
                }
            }
        },
        "/api/stock/queryStockData": {
            "post": {
                "description": "Return Strategy Sepcific Computed Data",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "Query Sepcific Stock Computed Data",
                "operationId": "queryStockData",
                "parameters": [
                    {
                        "description": "Query Stock List",
                        "name": "request_json",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.QueryStockDataRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.QueryStockDataResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "model.FactorRawData": {
            "type": "object",
            "properties": {
                "ema13": {
                    "type": "array",
                    "items": {
                        "type": "number"
                    }
                },
                "ema34": {
                    "type": "array",
                    "items": {
                        "type": "number"
                    }
                },
                "ema55": {
                    "type": "array",
                    "items": {
                        "type": "number"
                    }
                },
                "ma13": {
                    "type": "array",
                    "items": {
                        "type": "number"
                    }
                },
                "ma34": {
                    "type": "array",
                    "items": {
                        "type": "number"
                    }
                },
                "ma55": {
                    "type": "array",
                    "items": {
                        "type": "number"
                    }
                }
            }
        },
...
        "model.GetRecommandStockRequest": {
            "type": "object",
            "properties": {
                "query_data_len": {
                    "type": "integer"
                },
                "query_date": {
                    "type": "string",
                    "example": "2021-07-01"
                }
            }
        },
        "model.GetRecommandStockResponse": {
            "type": "object",
            "properties": {
                "stock_datas": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/model.SingleStockData"
                    }
                },
                "stock_prediction_datas": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/model.StockPredictItem"
                    }
                }
            }
        },
        "model.GetTotalMarketIndexDataRequest": {
            "type": "object",
            "properties": {
                "query_date": {
                    "type": "string",
                    "example": "2021-07-01"
                }
            }
        },
        "model.GetTotalMarketIndexDataResponse": {
            "type": "object",
            "properties": {
                "index_pred_data": {
                    "$ref": "#/definitions/model.SinglePredictRecord"
                },
                "index_raw_data": {
                    "$ref": "#/definitions/model.SinglePredictRecord"
                },
                "show_msg": {
                    "type": "string"
                }
            }
        },
...
        "model.MarketRawData": {
            "type": "object",
            "properties": {
                "close": {
                    "type": "array",
                    "items": {
                        "type": "number"
                    }
                },
                "date": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "high": {
                    "type": "array",
                    "items": {
                        "type": "number"
                    }
                },
                "low": {
                    "type": "array",
                    "items": {
                        "type": "number"
                    }
                },
                "money": {
                    "type": "array",
                    "items": {
                        "type": "number"
                    }
                },
                "open": {
                    "type": "array",
                    "items": {
                        "type": "number"
                    }
                },
                "volume": {
                    "type": "array",
                    "items": {
                        "type": "number"
                    }
                }
            }
        },
        "model.QueryStockDataRequest": {
            "type": "object",
            "properties": {
                "query_data_len": {
                    "type": "integer"
                },
                "stock_list": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                }
            }
        },
        "model.QueryStockDataResponse": {
            "type": "object",
            "properties": {
                "stock_datas": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/model.SingleStockData"
                    }
                }
            }
        },
        "model.SinglePredictRecord": {
            "type": "object",
            "properties": {
                "close": {
                    "type": "array",
                    "items": {
                        "type": "number"
                    }
                },
                "date": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "high": {
                    "type": "array",
                    "items": {
                        "type": "number"
                    }
                },
                "low": {
                    "type": "array",
                    "items": {
                        "type": "number"
                    }
                },
                "open": {
                    "type": "array",
                    "items": {
                        "type": "number"
                    }
                }
            }
        },
        "model.SingleStockData": {
            "type": "object",
            "properties": {
                "stock_info": {
                    "$ref": "#/definitions/model.StockBasicInfo"
                },
                "stock_raw_datas": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/model.StockSinglePeriodData"
                    }
                }
            }
        },
        "model.StockBasicInfo": {
            "type": "object",
            "properties": {
                "exchange": {
                    "type": "string",
                    "example": "BJSE/SZSE/SSE"
                },
                "stock_code": {
                    "type": "string",
                    "example": "002142.SZ"
                },
                "stock_industry": {
                    "type": "string",
                    "example": "石油"
                },
                "stock_name": {
                    "type": "string",
                    "example": "宁波银行"
                }
            }
        },
        "model.StockPredictItem": {
            "type": "object",
            "properties": {
                "prediction_record": {
                    "$ref": "#/definitions/model.SinglePredictRecord"
                },
                "show_msg": {
                    "type": "string"
                },
                "stock_info": {
                    "$ref": "#/definitions/model.StockBasicInfo"
                }
            }
        },
        "model.StockSinglePeriodData": {
            "type": "object",
            "properties": {
                "factor_raw_data": {
                    "$ref": "#/definitions/model.FactorRawData"
                },
                "market_raw_data": {
                    "$ref": "#/definitions/model.MarketRawData"
                },
                "period_type": {
                    "type": "string"
                }
            }
        }
    }
}